chore: moved Pokt logic to context
