+++ src/types/environment.d.ts
@@ -8,5 +8,9 @@ declare global {
       SECURE_API_KEY: string;
     }
   }
+
+  interface Window {
+    pocketNetwork: any;
+  }
 }
 export {};
